- implemented partial support for multiple task lists - added preview of subtasks in the task list (unfinished) - minor corrections in the code

+++ app/src/main/java/com/example/taskorganiserapp/SubtaskItemViewHolder.kt
@@ -2,6 +2,7 @@ package com.example.taskorganiserapp
 
 import android.content.Context
 import android.graphics.Paint
+import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.example.taskorganiserapp.databinding.SubtaskItemBinding
 import com.example.taskorganiserapp.databinding.TaskItemBinding
@@ -29,8 +30,13 @@ class SubtaskItemViewHolder (
                 clickListener.setIncompleteSubtaskItem(subtask)
         }
 
-        binding.deleteSubtask.setOnClickListener {
-            clickListener.deleteSubtaskItem(subtask)
+        if(subtask.creatorMode) {
+            binding.deleteSubtask.visibility = View.VISIBLE
+            binding.deleteSubtask.setOnClickListener {
+                clickListener.deleteSubtaskItem(subtask)
+            }
         }
+        else
+            binding.deleteSubtask.visibility = View.GONE
     }
 }
